+++ shared/subgraph_generator.cu
@@ -3,14 +3,14 @@
 #include "subgraph.cuh"
 #include "gpu_error_check.cuh"
 
-const unsigned int NUM_THREADS = 64;
+const size_t NUM_THREADS = 64;
 
-const unsigned int THRESHOLD_THREAD = 50000;
+const size_t THRESHOLD_THREAD = 50000;
 
-__global__ void prePrefix(unsigned int *activeNodesLabeling, unsigned int *activeNodesDegree, 
-							unsigned int *outDegree, bool *label1, bool *label2, unsigned int numNodes)
+__global__ void prePrefix(unsigned int *activeNodesLabeling, size_t *activeNodesDegree, 
+							unsigned int *outDegree, bool *label1, bool *label2, size_t numNodes)
 {
-	unsigned int id = blockDim.x * blockIdx.x + threadIdx.x;
+	size_t id = blockDim.x * blockIdx.x + threadIdx.x;
 	if(id < numNodes){
 		activeNodesLabeling[id] = label1[id] || label2[id]; // label1 is always zero in sync
 		//activeNodesLabeling[id] = label[id];
@@ -21,10 +21,10 @@ __global__ void prePrefix(unsigned int *activeNodesLabeling, unsigned int *activ
 	}	
 }
 
-__global__ void prePrefix(unsigned int *activeNodesLabeling, unsigned int *activeNodesDegree, 
-							unsigned int *outDegree, float *delta, unsigned int numNodes, float acc)
+__global__ void prePrefix(unsigned int *activeNodesLabeling, size_t *activeNodesDegree, 
+							unsigned int *outDegree, float *delta, size_t numNodes, float acc)
 {
-	unsigned int id = blockDim.x * blockIdx.x + threadIdx.x;
+	size_t id = blockDim.x * blockIdx.x + threadIdx.x;
 	if(id < numNodes){
 		if(delta[id] > acc)
 		{
@@ -40,20 +40,20 @@ __global__ void prePrefix(unsigned int *activeNodesLabeling, unsigned int *activ
 	}
 }
 
-__global__ void makeQueue(unsigned int *activeNodes, unsigned int *activeNodesLabeling,
-							unsigned int *prefixLabeling, unsigned int numNodes)
+__global__ void makeQueue(size_t *activeNodes, unsigned int *activeNodesLabeling,
+							size_t *prefixLabeling, size_t numNodes)
 {
-	unsigned int id = blockDim.x * blockIdx.x + threadIdx.x;
+	size_t id = blockDim.x * blockIdx.x + threadIdx.x;
 	if(id < numNodes && activeNodesLabeling[id] == 1){
 		activeNodes[prefixLabeling[id]] = id;
 	}
 }
 
-__global__ void makeActiveNodesPointer(unsigned int *activeNodesPointer, unsigned int *activeNodesLabeling, 
-											unsigned int *prefixLabeling, unsigned int *prefixSumDegrees, 
-											unsigned int numNodes)
+__global__ void makeActiveNodesPointer(size_t *activeNodesPointer, unsigned int *activeNodesLabeling, 
+											size_t *prefixLabeling, size_t *prefixSumDegrees, 
+											size_t numNodes)
 {
-	unsigned int id = blockDim.x * blockIdx.x + threadIdx.x;
+	size_t id = blockDim.x * blockIdx.x + threadIdx.x;
 	if(id < numNodes && activeNodesLabeling[id] == 1){
 		activeNodesPointer[prefixLabeling[id]] = prefixSumDegrees[id];
 	}
@@ -61,34 +61,34 @@ __global__ void makeActiveNodesPointer(unsigned int *activeNodesPointer, unsigne
 
 // pthread
 template <class E>
-void dynamic(unsigned int tId,
-				unsigned int numThreads,	
-				unsigned int numActiveNodes,
-				unsigned int *activeNodes,
+void dynamic(size_t tId,
+				size_t numThreads,	
+				size_t numActiveNodes,
+				size_t *activeNodes,
 				unsigned int *outDegree, 
-				unsigned int *activeNodesPointer,
-				unsigned int *nodePointer, 
+				size_t *activeNodesPointer,
+				size_t *nodePointer, 
 				E *activeEdgeList,
 				E *edgeList)
 {
 
-	unsigned int chunkSize = ceil(numActiveNodes / numThreads);
-	unsigned int left, right;
+	size_t chunkSize = ceil(numActiveNodes / numThreads);
+	size_t left, right;
 	left = tId * chunkSize;
 	right = min(left+chunkSize, numActiveNodes);	
 	
-	unsigned int thisNode;
-	unsigned int thisDegree;
-	unsigned int fromHere;
-	unsigned int fromThere;
+	size_t thisNode;
+	size_t thisDegree;
+	size_t fromHere;
+	size_t fromThere;
 
-	for(unsigned int i=left; i<right; i++)
+	for(size_t i=left; i<right; i++)
 	{
 		thisNode = activeNodes[i];
 		thisDegree = outDegree[thisNode];
 		fromHere = activeNodesPointer[i];
 		fromThere = nodePointer[thisNode];
-		for(unsigned int j=0; j<thisDegree; j++)
+		for(size_t j=0; j<thisDegree; j++)
 		{
 			activeEdgeList[fromHere+j] = edgeList[fromThere+j];
 		}
@@ -100,28 +100,28 @@ template <class E>
 SubgraphGenerator<E>::SubgraphGenerator(Graph<E> &graph)
 {
 	gpuErrorcheck(cudaMallocHost(&activeNodesLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&activeNodesDegree, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&prefixLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&prefixSumDegrees, (graph.num_nodes+1) * sizeof(unsigned int)));
+	gpuErrorcheck(cudaMallocHost(&activeNodesDegree, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMallocHost(&prefixLabeling, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMallocHost(&prefixSumDegrees, (graph.num_nodes+1) * sizeof(size_t)));
 	
 	gpuErrorcheck(cudaMalloc(&d_activeNodesLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_activeNodesDegree, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_prefixLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_prefixSumDegrees , (graph.num_nodes+1) * sizeof(unsigned int)));
+	gpuErrorcheck(cudaMalloc(&d_activeNodesDegree, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMalloc(&d_prefixLabeling, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMalloc(&d_prefixSumDegrees , (graph.num_nodes+1) * sizeof(size_t)));
 }
 
 template <class E>
 SubgraphGenerator<E>::SubgraphGenerator(GraphPR<E> &graph)
 {
 	gpuErrorcheck(cudaMallocHost(&activeNodesLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&activeNodesDegree, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&prefixLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMallocHost(&prefixSumDegrees, (graph.num_nodes+1) * sizeof(unsigned int)));
+	gpuErrorcheck(cudaMallocHost(&activeNodesDegree, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMallocHost(&prefixLabeling, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMallocHost(&prefixSumDegrees, (graph.num_nodes+1) * sizeof(size_t)));
 	
 	gpuErrorcheck(cudaMalloc(&d_activeNodesLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_activeNodesDegree, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_prefixLabeling, graph.num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_prefixSumDegrees , (graph.num_nodes+1) * sizeof(unsigned int)));
+	gpuErrorcheck(cudaMalloc(&d_activeNodesDegree, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMalloc(&d_prefixLabeling, graph.num_nodes * sizeof(size_t)));
+	gpuErrorcheck(cudaMalloc(&d_prefixSumDegrees , (graph.num_nodes+1) * sizeof(size_t)));
 }
 
 template <class E>
@@ -133,7 +133,7 @@ void SubgraphGenerator<E>::generate(Graph<E> &graph, Subgraph<E> &subgraph)
 	prePrefix<<<graph.num_nodes/512+1, 512>>>(d_activeNodesLabeling, d_activeNodesDegree, graph.d_outDegree, graph.d_label1, graph.d_label2, graph.num_nodes);
 		
 	thrust::device_ptr<unsigned int> ptr_labeling(d_activeNodesLabeling);
-	thrust::device_ptr<unsigned int> ptr_labeling_prefixsum(d_prefixLabeling);
+	thrust::device_ptr<size_t> ptr_labeling_prefixsum(d_prefixLabeling);
 	
 	subgraph.numActiveNodes = thrust::reduce(ptr_labeling, ptr_labeling + graph.num_nodes);
 	//cout << "Number of Active Nodes = " << subgraph.numActiveNodes << endl;
@@ -142,24 +142,24 @@ void SubgraphGenerator<E>::generate(Graph<E> &graph, Subgraph<E> &subgraph)
 	
 	makeQueue<<<graph.num_nodes/512+1, 512>>>(subgraph.d_activeNodes, d_activeNodesLabeling, d_prefixLabeling, graph.num_nodes);
 	
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodes, subgraph.d_activeNodes, subgraph.numActiveNodes*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodes, subgraph.d_activeNodes, subgraph.numActiveNodes*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
-	thrust::device_ptr<unsigned int> ptr_degrees(d_activeNodesDegree);
-	thrust::device_ptr<unsigned int> ptr_degrees_prefixsum(d_prefixSumDegrees);
+	thrust::device_ptr<size_t> ptr_degrees(d_activeNodesDegree);
+	thrust::device_ptr<size_t> ptr_degrees_prefixsum(d_prefixSumDegrees);
 	
 	thrust::exclusive_scan(ptr_degrees, ptr_degrees + graph.num_nodes, ptr_degrees_prefixsum);
 	
 	makeActiveNodesPointer<<<graph.num_nodes/512+1, 512>>>(subgraph.d_activeNodesPointer, d_activeNodesLabeling, d_prefixLabeling, d_prefixSumDegrees, graph.num_nodes);
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, subgraph.numActiveNodes*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, subgraph.numActiveNodes*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
-	unsigned int numActiveEdges = 0;
+	size_t numActiveEdges = 0;
 	if(subgraph.numActiveNodes>0)
 		numActiveEdges = subgraph.activeNodesPointer[subgraph.numActiveNodes-1] + graph.outDegree[subgraph.activeNodes[subgraph.numActiveNodes-1]];	
 	
-	unsigned int last = numActiveEdges;
-	gpuErrorcheck(cudaMemcpy(subgraph.d_activeNodesPointer+subgraph.numActiveNodes, &last, sizeof(unsigned int), cudaMemcpyHostToDevice));
+	size_t last = numActiveEdges;
+	gpuErrorcheck(cudaMemcpy(subgraph.d_activeNodesPointer+subgraph.numActiveNodes, &last, sizeof(size_t), cudaMemcpyHostToDevice));
 	
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, (subgraph.numActiveNodes+1)*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, (subgraph.numActiveNodes+1)*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
 	
 	//finishDynG = std::chrono::system_clock::now();
@@ -170,14 +170,14 @@ void SubgraphGenerator<E>::generate(Graph<E> &graph, Subgraph<E> &subgraph)
 	//td::chrono::time_point<std::chrono::system_clock> startDynC, finishDynC;
 	//startDynC = std::chrono::system_clock::now();
 	
-	unsigned int numThreads = NUM_THREADS;
+	size_t numThreads = NUM_THREADS;
 
 	if(subgraph.numActiveNodes < THRESHOLD_THREAD)
 		numThreads = 1;
 
 	thread runThreads[numThreads];
 	
-	for(unsigned int t=0; t<numThreads; t++)
+	for(size_t t=0; t<numThreads; t++)
 	{
 
 		runThreads[t] = thread(dynamic<E>,
@@ -193,7 +193,7 @@ void SubgraphGenerator<E>::generate(Graph<E> &graph, Subgraph<E> &subgraph)
 
 	}
 		
-	for(unsigned int t=0; t<numThreads; t++)
+	for(size_t t=0; t<numThreads; t++)
 		runThreads[t].join();
 	
 	//finishDynC = std::chrono::system_clock::now();
@@ -214,7 +214,7 @@ void SubgraphGenerator<E>::generate(GraphPR<E> &graph, Subgraph<E> &subgraph, fl
 	prePrefix<<<graph.num_nodes/512+1, 512>>>(d_activeNodesLabeling, d_activeNodesDegree, graph.d_outDegree, graph.d_delta, graph.num_nodes, acc);
 		
 	thrust::device_ptr<unsigned int> ptr_labeling(d_activeNodesLabeling);
-	thrust::device_ptr<unsigned int> ptr_labeling_prefixsum(d_prefixLabeling);
+	thrust::device_ptr<size_t> ptr_labeling_prefixsum(d_prefixLabeling);
 	
 	subgraph.numActiveNodes = thrust::reduce(ptr_labeling, ptr_labeling + graph.num_nodes);
 	//cout << "Number of Active Nodes = " << subgraph.numActiveNodes << endl;
@@ -223,24 +223,24 @@ void SubgraphGenerator<E>::generate(GraphPR<E> &graph, Subgraph<E> &subgraph, fl
 	
 	makeQueue<<<graph.num_nodes/512+1, 512>>>(subgraph.d_activeNodes, d_activeNodesLabeling, d_prefixLabeling, graph.num_nodes);
 	
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodes, subgraph.d_activeNodes, subgraph.numActiveNodes*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodes, subgraph.d_activeNodes, subgraph.numActiveNodes*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
-	thrust::device_ptr<unsigned int> ptr_degrees(d_activeNodesDegree);
-	thrust::device_ptr<unsigned int> ptr_degrees_prefixsum(d_prefixSumDegrees);
+	thrust::device_ptr<size_t> ptr_degrees(d_activeNodesDegree);
+	thrust::device_ptr<size_t> ptr_degrees_prefixsum(d_prefixSumDegrees);
 	
 	thrust::exclusive_scan(ptr_degrees, ptr_degrees + graph.num_nodes, ptr_degrees_prefixsum);
 	
 	makeActiveNodesPointer<<<graph.num_nodes/512+1, 512>>>(subgraph.d_activeNodesPointer, d_activeNodesLabeling, d_prefixLabeling, d_prefixSumDegrees, graph.num_nodes);
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, subgraph.numActiveNodes*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, subgraph.numActiveNodes*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
-	unsigned int numActiveEdges = 0;
+	size_t numActiveEdges = 0;
 	if(subgraph.numActiveNodes>0)
 		numActiveEdges = subgraph.activeNodesPointer[subgraph.numActiveNodes-1] + graph.outDegree[subgraph.activeNodes[subgraph.numActiveNodes-1]];	
 	
-	unsigned int last = numActiveEdges;
-	gpuErrorcheck(cudaMemcpy(subgraph.d_activeNodesPointer+subgraph.numActiveNodes, &last, sizeof(unsigned int), cudaMemcpyHostToDevice));
+	size_t last = numActiveEdges;
+	gpuErrorcheck(cudaMemcpy(subgraph.d_activeNodesPointer+subgraph.numActiveNodes, &last, sizeof(size_t), cudaMemcpyHostToDevice));
 	
-	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, (subgraph.numActiveNodes+1)*sizeof(unsigned int), cudaMemcpyDeviceToHost));
+	gpuErrorcheck(cudaMemcpy(subgraph.activeNodesPointer, subgraph.d_activeNodesPointer, (subgraph.numActiveNodes+1)*sizeof(size_t), cudaMemcpyDeviceToHost));
 	
 	
 	//finishDynG = std::chrono::system_clock::now();
@@ -251,14 +251,14 @@ void SubgraphGenerator<E>::generate(GraphPR<E> &graph, Subgraph<E> &subgraph, fl
 	//td::chrono::time_point<std::chrono::system_clock> startDynC, finishDynC;
 	//startDynC = std::chrono::system_clock::now();
 	
-	unsigned int numThreads = NUM_THREADS;
+	size_t numThreads = NUM_THREADS;
 
 	if(subgraph.numActiveNodes < THRESHOLD_THREAD)
 		numThreads = 1;
 
 	thread runThreads[numThreads];
 	
-	for(unsigned int t=0; t<numThreads; t++)
+	for(size_t t=0; t<numThreads; t++)
 	{
 
 		runThreads[t] = thread(dynamic<E>,
@@ -274,7 +274,7 @@ void SubgraphGenerator<E>::generate(GraphPR<E> &graph, Subgraph<E> &subgraph, fl
 
 	}
 		
-	for(unsigned int t=0; t<numThreads; t++)
+	for(size_t t=0; t<numThreads; t++)
 		runThreads[t].join();
 	
 	//finishDynC = std::chrono::system_clock::now();
